+++ runtime/master/pom.xml
@@ -92,17 +92,5 @@ under the License.
       <version>1.6.2</version>
       <scope>test</scope>
     </dependency>
-    <dependency>
-      <groupId>com.amazonaws</groupId>
-      <artifactId>aws-java-sdk-core</artifactId>
-      <version>1.11.429</version>
-      <scope>compile</scope>
-    </dependency>
-    <dependency>
-      <groupId>com.amazonaws</groupId>
-      <artifactId>aws-java-sdk-lambda</artifactId>
-      <version>1.11.429</version>
-      <scope>compile</scope>
-    </dependency>
   </dependencies>
 </project>
