+++ runtime/master/src/main/java/org/apache/nemo/runtime/master/scheduler/StreamingScheduler.java
@@ -150,10 +150,11 @@ public final class StreamingScheduler implements Scheduler {
   }
 
   @Override
-  public void onExecutorAdded(final ExecutorRepresenter executorRepresenter) {
-    LOG.info("{} added (node: {})", executorRepresenter.getExecutorId(), executorRepresenter.getNodeName());
+  public void onExecutorAdded(final ExecutorRepresenter defaultExecutorRepresenter) {
+    LOG.info("{} added (node: {})", defaultExecutorRepresenter.getExecutorId(),
+      defaultExecutorRepresenter.getNodeName());
     taskDispatcher.onExecutorSlotAvailable();
-    executorRegistry.registerExecutor(executorRepresenter);
+    executorRegistry.registerExecutor(defaultExecutorRepresenter);
   }
 
   @Override
