+++ runtime/master/src/test/java/org/apache/nemo/runtime/master/scheduler/AntiAffinitySchedulingConstraintTest.java
@@ -24,7 +24,7 @@ import org.apache.nemo.common.ir.vertex.executionproperty.ResourceAntiAffinityPr
 import org.apache.nemo.runtime.common.RuntimeIdManager;
 import org.apache.nemo.runtime.common.plan.Stage;
 import org.apache.nemo.runtime.common.plan.Task;
-import org.apache.nemo.runtime.master.resource.ExecutorRepresenter;
+import org.apache.nemo.runtime.master.resource.DefaultExecutorRepresenter;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.powermock.core.classloader.annotations.PrepareForTest;
@@ -42,7 +42,7 @@ import static org.mockito.Mockito.when;
  * Test cases for {@link AntiAffinitySchedulingConstraint}.
  */
 @RunWith(PowerMockRunner.class)
-@PrepareForTest({ExecutorRepresenter.class, Task.class, Stage.class, IRVertex.class})
+@PrepareForTest({DefaultExecutorRepresenter.class, Task.class, Stage.class, IRVertex.class})
 public final class AntiAffinitySchedulingConstraintTest {
   private final static int FIRST_ATTEMPT = 0;
   private final static String EXEc_MAP_ID = "ID";
@@ -54,12 +54,12 @@ public final class AntiAffinitySchedulingConstraintTest {
     return task;
   }
 
-  private static ExecutorRepresenter mockExecutorRepresenter(final Task task) {
-    final ExecutorRepresenter executorRepresenter = mock(ExecutorRepresenter.class);
+  private static DefaultExecutorRepresenter mockExecutorRepresenter(final Task task) {
+    final DefaultExecutorRepresenter defaultExecutorRepresenter = mock(DefaultExecutorRepresenter.class);
     final Set<Task> runningTasks = new HashSet<>();
     runningTasks.add(task);
-    when(executorRepresenter.getRunningTasks()).thenReturn(runningTasks);
-    return executorRepresenter;
+    when(defaultExecutorRepresenter.getRunningTasks()).thenReturn(runningTasks);
+    return defaultExecutorRepresenter;
   }
 
   /**
@@ -78,7 +78,7 @@ public final class AntiAffinitySchedulingConstraintTest {
     final Task task0 = mockTask(0, emap);  // anti-affinity task
     final Task task1 = mockTask(1, emap);  // anti-affinity task
     final Task task2 = mockTask(2, emap);  // normal task
-    final ExecutorRepresenter e0 = mockExecutorRepresenter(task0);  // schedule task0 to e0
+    final DefaultExecutorRepresenter e0 = mockExecutorRepresenter(task0);  // schedule task0 to e0
 
     assertEquals(false, schedulingConstraint.testSchedulability(e0, task1));
     assertEquals(true, schedulingConstraint.testSchedulability(e0, task2));
