+++ runtime/pom.xml
@@ -37,7 +37,6 @@ under the License.
     <module>executor</module>
     <module>master</module>
     <module>test</module>
-    <module>lambda</module>
   </modules>
 
 </project>
